+++ Core/BlazorMvvm.Core/Commands/BlazorAsyncCommand.cs
@@ -6,6 +6,7 @@ namespace BlazorMvvm;
 public class BlazorAsyncCommand : IBlazorAsyncCommand
 {
     private readonly SemaphoreSlim _executionSemaphore = new(1, 1);
+    private int _runningTasksCount;
 
     private readonly Func<Task> _execute;
     private readonly Func<Task<bool>>? _canExecute;
@@ -25,6 +26,8 @@ public class BlazorAsyncCommand : IBlazorAsyncCommand
         }
     }
 
+    public bool ContinueOnCapturedContext { get; set; } = true;
+
     public BlazorAsyncCommand(Func<Task> execute, Func<Task<bool>>? canExecute = null, bool allowConcurrentExecutions = false)
     {
         this._execute = execute ?? throw new ArgumentNullException(nameof(execute));
@@ -35,24 +38,35 @@ public class BlazorAsyncCommand : IBlazorAsyncCommand
     public async Task<bool> CanExecute()
     {
         if (!AllowConcurrentExecutions && IsExecuting) return false;
-        return this._canExecute == null || await this._canExecute();
+        return this._canExecute == null || await this._canExecute().ConfigureAwait(ContinueOnCapturedContext);
     }
 
-    public async void Execute()
+    public async Task ExecuteAsync()
     {
-        if (this._canExecute != null && !await this._canExecute()) return;
-        bool shouldExecute = AllowConcurrentExecutions || await _executionSemaphore.WaitAsync(0);
+        if (this._canExecute != null && !await this._canExecute().ConfigureAwait(ContinueOnCapturedContext)) return;
+        bool shouldExecute = AllowConcurrentExecutions || await _executionSemaphore.WaitAsync(0).ConfigureAwait(ContinueOnCapturedContext);
         if (!shouldExecute) return;
 
         try
         {
-            IsExecuting = true;
-            await this._execute();
+            if (Interlocked.Increment(ref _runningTasksCount) == 1)
+            {
+                IsExecuting = true;
+            }
+            await this._execute().ConfigureAwait(ContinueOnCapturedContext);
         }
         finally
         {
-            IsExecuting = false;
+            if (Interlocked.Decrement(ref _runningTasksCount) == 0)
+            {
+                IsExecuting = false;
+            }
             if (!AllowConcurrentExecutions) _executionSemaphore.Release();
         }
+    }
+
+    public void Execute()
+    {
+        _ = ExecuteAsync();
     }
 }