+++ Core/BlazorMvvm.Core/Commands/Interfaces/IBlazorAsyncRelayCommandT.cs
@@ -6,7 +6,9 @@ public interface IBlazorAsyncRelayCommand<T>
     public delegate void IsExecutingChanged(bool isExecuting);
     public event IsExecutingChanged? OnIsExecutingChanged;
     public bool AllowConcurrentExecutions { get; set; }
+    public bool ContinueOnCapturedContext { get; set; }
     public bool IsExecuting { get; }
     public Task<bool> CanExecute(T arg);
+    public Task ExecuteAsync(T arg);
     public void Execute(T arg);
 }