+++ Tests/BlazorMvvm.Tests/ViewModel/Collision/NamespaceA/CollisionViewModel.cs
@@ -1,0 +1,9 @@
+﻿namespace BlazorMvvm.Tests.ViewModel.Collision.NamespaceA
+{
+    [BlazorMvvmViewModel]
+    public partial class CollisionViewModel : BlazorViewModel
+    {
+        [BlazorObservableProperty]
+        private string _valueA = "";
+    }
+}