+++ Tests/BlazorMvvm.Tests/ViewModel/Collision/NamespaceB/CollisionViewModel.cs
@@ -1,0 +1,9 @@
+﻿namespace BlazorMvvm.Tests.ViewModel.Collision.NamespaceB
+{
+    [BlazorMvvmViewModel]
+    public partial class CollisionViewModel : BlazorViewModel
+    {
+        [BlazorObservableProperty]
+        private string _valueB = "";
+    }
+}